Moved all files to the 'mints' directory.
