--- candies/cli/__init__.py
@@ -1,3 +1,0 @@
-from candies.cli.cli import CLI, cli
-from candies.cli.command import Command
-from candies.cli.args import Arg, Opt, Flag, Typed
