--- candies/cli/parsers/__init__.py
@@ -1,2 +1,0 @@
-from candies.cli.parsers.parser import Parser
-from candies.cli.parsers.standard import StandardParser
