+++ mints/__init__.py
@@ -1,0 +1,3 @@
+from mints.cli import CLI, cli
+from mints.command import Command
+from mints.args import Arg, Opt, Flag, Typed
