+++ mints/parsers/__init__.py
@@ -1,0 +1,2 @@
+from mints.parsers.parser import Parser
+from mints.parsers.standard import StandardParser
