+++ tests/integration/command_test.py
@@ -4,10 +4,10 @@ from typing import Any
 
 import pytest
 
-from candies.cli.args.arg import Arg
-from candies.cli.args.opt import Opt
-from candies.cli.args.flag import Flag
-from candies.cli.cli import cli, CLI
+from mints.args.arg import Arg
+from mints.args.opt import Opt
+from mints.args.flag import Flag
+from mints.cli import cli, CLI
 
 from tests.execution import execute
