+++ tests/integration/flag_test.py
@@ -5,9 +5,9 @@ from argparse import ArgumentError
 
 import pytest
 
-from candies.cli.cli import cli, CLI
-from candies.cli.args.flag import Flag
-from candies.cli.args.arg import Arg
+from mints.cli import cli, CLI
+from mints.args.flag import Flag
+from mints.args.arg import Arg
 
 from tests.execution import execute, redirect_stderr
