+++ tests/integration/typed_test.py
@@ -4,9 +4,9 @@ from typing import Any, List
 
 import pytest
 
-from candies.cli.args.arg import Arg
-from candies.cli.args.opt import Opt
-from candies.cli.cli import cli, CLI
+from mints.args.arg import Arg
+from mints.args.opt import Opt
+from mints.cli import cli, CLI
 
 from tests.execution import execute, redirect_stderr
